Fix blinking badge for non-English users, fixes #63
When someone messaged you when Messenger is in background, the title blinks between the following 2 forms:

- “(1) Messenger”
- “{name} messaged you”

In window-behaviour.js L111, it tries to parse out the “messaged you” from the title to prevent badge blink.
However for non-English users, this fails because the specified phrase does not exists in page title.

Therefore I'm trying the opposite way. If you cannot find the word “Messenger” in page title, I assume that
it's biinking for the unread message.